Fold structural parent capture back into its one consumer
The reverted structural-route replay left _StructuralParentCapture in
two_level.py with no user there; working_golden imported the private class
only to subclass it with a super()-only on_splice override. Restore the
capture as _ProposalLoopIdentity's own methods and drop two_level's unused
copy import. Also remove the durations.json entry for the reverted
test_placement_route_replays_from_a_cold_db.
