+++ tests/durations.json
@@ -349,7 +349,6 @@
  "tests/compiler/pipeline/search/test_two_level.py::test_inner_reward_shares_identical_kernel": 0.18,
  "tests/compiler/pipeline/search/test_two_level.py::test_node_rows_are_fold_ready": 0.25,
  "tests/compiler/pipeline/search/test_two_level.py::test_run_two_level_tune_single_terminal_assembles_bests": 2.12,
- "tests/compiler/pipeline/search/test_two_level_strategy.py::test_placement_route_replays_from_a_cold_db": 0.16,
  "tests/compiler/pipeline/search/test_two_level_strategy.py::test_scoring_is_separable_over_unique_kernels": 2.41,
  "tests/compiler/pipeline/test_knob.py::test_registry_complete_in_a_bare_process": 0.1,
  "tests/compiler/pipeline/test_resolve.py::test_decide_score_lands_on_trace": 0.11,
